+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- Add command line option to disable logging (#80)
+- Persist column widths across restarts (#53)
+- Add :clearcache command to clear cache files
+
 ## [1.1.0]
 
 ### Added